Move dataset_to_point_list to arviz module
