+++ tests/test_util.py
@@ -26,7 +26,6 @@ from pymc.distributions.transforms import Transform
 from pymc.util import (
     UNSET,
     _get_seeds_per_chain,
-    dataset_to_point_list,
     drop_warning_stat,
     get_value_vars_from_user_vars,
     hash_key,
@@ -154,38 +153,6 @@ def test_unset_repr(capsys):
     help(fn)
     captured = capsys.readouterr()
     assert "a=UNSET" in captured.out
-
-
-@pytest.mark.parametrize("input_type", ("dict", "Dataset"))
-def test_dataset_to_point_list(input_type):
-    if input_type == "dict":
-        ds = {}
-    elif input_type == "Dataset":
-        ds = xarray.Dataset()
-    ds["A"] = xarray.DataArray([[1, 2, 3]] * 2, dims=("chain", "draw"))
-    pl, _ = dataset_to_point_list(ds, sample_dims=["chain", "draw"])
-    assert isinstance(pl, list)
-    assert len(pl) == 6
-    assert isinstance(pl[0], dict)
-    assert isinstance(pl[0]["A"], np.ndarray)
-
-
-def test_transposed_dataset_to_point_list():
-    ds = xarray.Dataset()
-    ds["A"] = xarray.DataArray([[[1, 2, 3], [2, 3, 4]]] * 5, dims=("team", "draw", "chain"))
-    pl, _ = dataset_to_point_list(ds, sample_dims=["chain", "draw"])
-    assert isinstance(pl, list)
-    assert len(pl) == 6
-    assert isinstance(pl[0], dict)
-    assert isinstance(pl[0]["A"], np.ndarray)
-
-
-def test_dataset_to_point_list_str_key():
-    # Check that non-str keys are caught
-    ds = xarray.Dataset()
-    ds[3] = xarray.DataArray([1, 2, 3])
-    with pytest.raises(ValueError, match="must be str"):
-        dataset_to_point_list(ds, sample_dims=["chain", "draw"])
 
 
 def test_drop_warning_stat():
